refactor: refactor package structure
패키지 구조를 클린 아키텍처와 유사하게 수정
