+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".View.MapActivity"
+            android:name=".Presenter.View.MapActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -28,7 +28,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".View.PlaceSearchActivity"
+            android:name=".Presenter.View.PlaceSearchActivity"
             android:exported="false">
         </activity>
     </application>
