+++ app/src/main/java/campus/tech/kakao/map/Base/ViewModelFactory.kt
@@ -2,7 +2,7 @@ package campus.tech.kakao.map.Base
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import campus.tech.kakao.map.Repository.PlaceRepositoryImpl
+import campus.tech.kakao.map.Data.PlaceRepositoryImpl
 import campus.tech.kakao.map.ViewModel.SearchViewModel
 
 class ViewModelFactory(private val repository: PlaceRepositoryImpl) : ViewModelProvider.Factory {
