+++ app/src/main/java/campus/tech/kakao/map/Data/Mapper/EntityToModelMapper.kt
@@ -1,0 +1,5 @@
+package campus.tech.kakao.map.Data.Mapper
+
+interface EntityToModelMapper<Document,Place> {
+    fun map(document : Document) : Place
+}
