+++ app/src/main/java/campus/tech/kakao/map/Domain/Model/Place.kt
@@ -1,0 +1,3 @@
+package campus.tech.kakao.map.Domain.Model
+
+data class Place(val name: String, var address: String? = null, var category: PlaceCategory? = null)
