--- app/src/main/java/campus/tech/kakao/map/Mapper/EntityToModelMapper.kt
@@ -1,5 +1,0 @@
-package campus.tech.kakao.map.Mapper
-
-interface EntityToModelMapper<Document,Place> {
-    fun map(document : Document) : Place
-}
