--- app/src/main/java/campus/tech/kakao/map/Model/Place.kt
@@ -1,3 +1,0 @@
-package campus.tech.kakao.map.Model
-
-data class Place(val name: String, var address: String? = null, var category: PlaceCategory? = null)
