--- app/src/main/java/campus/tech/kakao/map/View/MapActivity.kt
@@ -1,47 +1,0 @@
-package campus.tech.kakao.map.View
-
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.os.Bundle
-import android.util.Log
-import android.widget.TextView
-import androidx.appcompat.app.AppCompatActivity
-import campus.tech.kakao.map.R
-import com.kakao.vectormap.KakaoMap
-import com.kakao.vectormap.KakaoMapReadyCallback
-import com.kakao.vectormap.KakaoMapSdk
-import com.kakao.vectormap.MapLifeCycleCallback
-import com.kakao.vectormap.MapView
-import java.security.MessageDigest
-
-
-class MapActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_map)
-
-        val mapView = findViewById<MapView>(R.id.mapView)
-        val searchText = findViewById<TextView>(R.id.search)
-
-        mapView.start(object : MapLifeCycleCallback() {
-            override fun onMapDestroy() {
-                // 지도 API 가 정상적으로 종료될 때 호출됨
-            }
-
-            override fun onMapError(error: Exception) {
-                // 인증 실패 및 지도 사용 중 에러가 발생할 때 호출됨
-
-            }
-        }, object : KakaoMapReadyCallback() {
-            override fun onMapReady(kakaoMap: KakaoMap) {
-                // 인증 후 API 가 정상적으로 실행될 때 호출됨
-            }
-        })
-
-        searchText.setOnClickListener {
-            val intent = Intent(this,PlaceSearchActivity::class.java)
-            startActivity(intent)
-        }
-
-    }
-}
