+++ app/src/main/java/campus/tech/kakao/map/di/AppContainer.kt
@@ -1,13 +1,13 @@
 package campus.tech.kakao.map.di
 
 import android.content.Context
-import campus.tech.kakao.map.Datasource.Local.Dao.FavoriteDaoImpl
-import campus.tech.kakao.map.Datasource.Local.Dao.PlaceDaoImpl
-import campus.tech.kakao.map.Datasource.Local.SqliteDB
-import campus.tech.kakao.map.Datasource.Remote.RetrofitService
-import campus.tech.kakao.map.Mapper.DocToPlaceMapper
-import campus.tech.kakao.map.Model.PlaceContract
-import campus.tech.kakao.map.Repository.PlaceRepositoryImpl
+import campus.tech.kakao.map.Data.Datasource.Local.Dao.FavoriteDaoImpl
+import campus.tech.kakao.map.Data.Datasource.Local.Dao.PlaceDaoImpl
+import campus.tech.kakao.map.Data.Datasource.Local.SqliteDB
+import campus.tech.kakao.map.Data.Datasource.Remote.RetrofitService
+import campus.tech.kakao.map.Data.Mapper.DocToPlaceMapper
+import campus.tech.kakao.map.Domain.Model.PlaceContract
+import campus.tech.kakao.map.Data.PlaceRepositoryImpl
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
