+++ app/src/main/res/layout/activity_map.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.MapActivity">
+    tools:context=".Presenter.View.MapActivity">
 
 
     <TextView
