+++ app/src/main/res/layout/activity_place_search.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.PlaceSearchActivity">
+    tools:context=".Presenter.View.PlaceSearchActivity">
 
     <EditText
         android:id="@+id/etSearchPlace"
